Handle errors during validation
In some cases, there may be errors in the incoming geometry
that cause JSTS operations to fail. For example, a self-
intersecting geometry fails during the CONUS validation step
as we try to intersect it with a CONUS boundary.

We now catch these errors and notify the user with a generic
message, instead of whitescreening on them.