Switch TR-55 Water Quality to Bullet Charts
Previously the water quality data was represented as a
horizontal bar chart, with one bar for each type of
observation. Unfortunately, the values of the three
observations was quite disparate in scale, making the
smallest bar insignificantly small.

To help bring each value to first class citizenship,
we switch to using three bullet charts. Each chart is
scaled to suit the value thereof, and thus shows real,
actionable information.