Do not pre-intersect streams with AoI
The longest part of the query was intersecting all the
streams with the area of interest before sending it to
the geoprocessing service, where each stream would be
intersected with the tile again for calculating the
results. Not only was this redundant, it was also the
primary source of latency in the mapshed submission.

By forgoing the extra intersection, and simply collecting
all the streams into a single geometry (MultiLineString),
the process is sped up by two orders of magnitude. Also,
since the intersection is already done in the following
step, the calculations do not change.

Furthermore, since the geoprocessing service is already
setup to handle an array of MultiLineStrings, by giving
it a singleton array of one MultiLineString, we can use
the geoprocessing service as is without needing a new
version.

For the sake of clarity, however, it would be good to
change that parameter to be a single MultiLineString
object rather than an array which conventionally has
a single item in it.

:zap: