Fix broken log configuration for geoprocessing service
The log routing configuration for the geoprocessing service via Beaver
was in a broken state, which was breaking all of the Beaver managed log
collection for the application as a whole. This change fixes the
geoprocessing configuration, which in turn enables all centralized
logging again (Nginx, application, tiler, Celery, Docker, etc.).