Ensure AWS profile is set for geoprocessing service
When the geoprocessing service is configured for local development,
ensure that the value of the Ansible `aws_profile` variable is set in
the service's Upstart configuration.

In the default case, `mmw-stg` is used, which should work fine because
the associated credentials are whitelisted for read access on the
DataHub AWS account.