Add Model, Serializer for HydroShareToken
The token response from HydroShare's OAuth server has
three important fields:

 - access_token:  To be used in all OAuth interactions
 - expires_in:    Seconds to expiry from when the token was
                  generated. Defaults to 30 days.
 - refresh_token: To be used to get a new access_token sans
                  user input.

When a token response is received from HydroShare, we save
it to the database linking it to the MMW user. All further
interactions can be done by looking up the token whenever
a user makes a request.

The serializer will likely not be used very often, since we
wouldn't want to expose the tokens. However, while this is
in development, it is a handy convenience for inspection.

:construction: :card_file_box: