Add HydroShareService module
In many ways similar to the ItsiService module, this extends
OAuth2Service and allows OAuth interaction. Furthermore, it
also interacts with the HydroShareToken models and can make
a HydroShare rest client, initialized to the right settings.

There are a few TODO items here, not implemented here for
they may be handled outside these functions in the views.
Since those views are still to be written, and this part
can possibly change, these items are left as comments.

:construction: :hankey: