Add HydroShare login endpoints
For authenticated users, we allow connecting their MMW account to
a HydroShare account. The retrieved credentials, which include an
access_token and refresh_token, are saved to the database.

This is facilitated via two endpoints which perform the OAuth
dance:

 - /user/hydroshare/login:
   Generates a redirect URI using the app credentials and sends
   the user there to login and give MMW read write permissions.

 - /user/hydroshare/authorize:
   Receives the authorization code from HydroShare and uses it
   to retrieve access tokens, then saves them to the database.

Currently the authorize endpoint also shows the saved token in an
API View for demonstration purposes, but this will be replaced by
a template that sends an HTML5 postMessage that the original page
which triggered the request is listening for.

:sparkles: