Add MultiShareModel view and template
This new share modal will be used to offer multiple sharing and
integration options to the user. Presently it only supports link
sharing, which can be turned on or off. When it is on, a textbox
with the link is shown with a copy button to copy it to the clip-
board.

Rather than take a modal-specific model, this expects the main
project model to be supplied.

This modal is expected to have more things, such as HydroShare
integration, in the near future. This also makes the "Make Public"
and "Make Private" buttons in the project options dropdown
obsolete. These should be removed in the future.

:sparkles: