Fix running MapShed on AoIs with no streams
Previously, if no streams intersected the area of interest,
instead of an array of stringified GeoJSON MultiLines, a
singleton array would be sent with the only value being
null. This caused parse issues in the geoprocessing service.

Now this case is caught on the Celery task itself, and if
so we return an empty object immediately, and do not invoke
the geoprocessing service at all. This short circuit makes
the feedback faster and removes unnecessary communication.

Refs #2518 :bug: