Calculate stream length for ag and non-ag areas
Reusing the nlcd_streams task from MapShed, we can get a
percentage value of how many streams are in agricultural
areas. Since this is done for all the streams in the area
of interest, and not by stream order, we get one value and
repeat it for all stream orders.

On the client side, this value is taken from the first
result (since they all have the same value), and combined
with the total length to get the total length in agricultural
areas. This is then subtracted from the total length to get
the total length in non-agricultural areas.

Refs #2528