Add Avg Slope To Streams Analysis Payload
We left a spot for the slope as `slopepct`.
I've renamed it to `avgslope` to keep it
as close to NHD name/treatment as possible
for API users. In the UI we still display
it as a percent.

I've also switched the empty treatment to
show "No Data" if there's no value for
a stream order. While "0 length" means
what it should, I think "0% slope" implies
there was some slope calculated when there's
just no data.