Add IFrame Modal view
This modal will contain an iframe within it. The iframe is
expected to fire a postMessage. The contents of the postMessage
are configured in the model given to this view. There are two
types of expected messages: a successSignal and a failureSignal.
If the modal receives a successSignal, it fires a success event
and closes itself in 0.5s. If it receives a failureSignal, it
fires a failure event. If it is closed before succeeding, it
fires a dismiss event.