Add HydroShare option, behavior
Adds a HydroShare slider to the Multi Share Modal. When the slider
is turned on, we check if the user has already connected their
account to HydroShare. If so, we simply let them slide it on or
off. If they haven't, they are shown a confirmation prompt that
asks them to connect to HydroShare. If they agree to it, they are
then taken to the HydroShare interface inside an iframe. Here
they can login and then allow access to model my watershed. If
the allowance succeeds, the iframe modal is closed and the slider
is updated. We re-fetch user details at this point to ensure that
any further changes to the slider do not re-trigger a login.