Use beta.hydroshare instead of dev.
Apparently dev is out of date and beta is a better reflection of
the current state of production. Based off of this comment:

https://github.com/hydroshare/hydroshare/issues/2530#issuecomment-352534078

From cursory observation, it seems like beta uses a clone of the
production database, while dev is its own environment. Thus, beta
uses the same OAuth credentials as production, instead of its own
set like dev.