Add HydroShareResource model, serializer, migration
When we export a project to HydroShare, we need to keep track
of its Resource id, title, date last exported, URL, and whether
or not to autosync. This, along with standard timestamps, is
tracked in a new HydroShareResource model, which has a 1-to-1
correspondence with the Project model. A Project can have 0 or
1 HydroShareResource, but a HydroShareResource belongs to
exactly 1 Project.

In HydroShare, the Resource ID (corresponding to the `resource`
field here) is a UUID. However, we store it as a Char(63) to
allow for possible changes in the HydroShare platform, as well
as to maintain representational consistency. HydroShare uses
small case, no-hyphen format for them, whereas Postgres and
Django use the small case, hyphenated format for UUIDs.

The `title` will be populated when exporting for the first time.
If it is changed in HydroShare afterwards, we will not know about
it since there is no hook in HydroShare to listen to.

The `url` property is dynamically generated using the configured
HYDROSHARE_BASE_URL and the resource ID.