Add HydroShare Export backend
Add a new app and top-level url family for housing exports.
So far we have only one: HydroShare, available at
`/export/hydroshare`, but if there are more in the future
they can all be grouped under `/export/`.

Currently we only support exporting saved Projects to HydroShare,
thus the Project must be specified in the call as so:
`/export/hydroshare?project=XX`. If this project does not
exist or does not belong to the user, we return a 404.

The endpoint supports GET, POST, and DELETE. GET simply returns
the HydroShareResource object corresponding to the project if
one exists.

POST for a project that already has an export currently behaves
the same as GET. This will be amended in the future to update
the HydroShareResource and re-export to HydroShare.

POST for a project that has not been exported creates a resource
in HydroShare with the given title, abstract, keywords, and any
supplied files. The area of interest of the project is saved as
a GeoJSON and as a Shapefile in HydroShare. Any additional files
coming from the client as {name: "string", contents: "string"}
objects are also added. A HydroShareResource object is created
for this project and returned.

DELETE removes the resource from HydroShare and also the matching
HydroShareResource object from MMW.