Add HydroShare Export Modal
This will be triggered from the MultiShare Modal before
exporting to HydroShare. This modal will allow the user
to pick a title (auto-populated with the Project's name),
write an abstract (required), and supply some optional
keywords.

If the user cancels out of this dialog, we do nothing.
Otherwise we trigger an `export` method with the values
of the various fields, to be used by the calling Multi
Share Modal.