Refactor HydroShare export into Project model
Previously the mechanics of the export were too tightly coupled
with the MultiShare Modal View. This made it difficult to reuse
the functionality elsewhere, and was brittle (e.g. breaking if
the modal was closed before an AJAX request completed).

By splitting the network requests into the project model and
updating the view only via renders, we make the code clean and
separate, and allow the reuse of HydroShare export in other
parts (such as after save).

This also ensures that if something else triggers a HydroShare
export (such as a project save), the share modal will behave
correctly (showing the spinner, switch state, etc).

This is a pure refactoring and doesn't change user-facing
behavior at all.