Allow setting / unsetting autosync
We add a PATCH handler to the export/hydroshare endpoint which
takes a request like {"autosync": false} and updates the database
with the new setting and sends back the updated HydroShareResource
object.

If the project is not found, or it isn't syncing to HydroShare,
we return a 404. If "autosync" isn't present in the request body,
we return a 400.