Add api, referrer fields to RequestLog
The api field will be True for direct API access via Token
Authentication, and False for UI access via Session
Authentication.

We default it to True because so far only the API requests have
been logged, so all existing rows will have correct values.

The referrer field will capture the source of the request, so
that some referrers may be excluded during reporting. This
will possibly be NULL for API calls.