Move share button to top level
Since now this is a very feature rich interface expected
to be used more often, we make it more prominent in the
user interface. The rules for its visibility are the same
as before, so only users that own the project will see it
after it has been saved once.

Also, add an `is-exporting` class to it when an export is
currently active. This will be styled later. Currently it
is only triggered by HydroShare export, but theoretically
could be similarly used by other integrations as well.