Show Warning Icon If Hydroshare Action Fails
If a hydroshare export request fails, show
a warning icon in the modal with a title
explaining the error. Because we cannot
guarentee that a 404 from Hydroshare always
means the resource has been deleted on
Hydroshare, we let the user decide if
they'd like to unsync the resource or
not.

Similar to our approach with `is_exporting`,
add an `error` class to the Share button
in the project menu for styling at a later
date.