Disconnect from HydroShare
When a user disconnects their account from HydroShare, we delete
their token, and all the HydroShareResource objects corresponding
to their projects.

If they ever re-connect to HydroShare, those projects will have
to be exported again, and they will export to all new resources,
not the ones they were previously connected to.

The resources on HydroShare will continue to exist, orphaned.