Add keyword bubble styling and behavior
Keywords can now be added from a text box and "Add keyword"
button, which will make a small bubble and add to the bottom
of the modal. These can be deleted with a little cross on each.

Behavior has been updated so that pressing "Enter" in the
keyword text box adds the keyword instead of submitting the
modal. Empty keywords are ignored.