Don't re-export Analyze files
Previously we would always export all files coming from
the front-end. However, this is really slow and leads to
frequent timeouts on the front-end. To alleviate this,
we check if the analyze files have been exported already.
If so, we exclude them.

This allows for the possibility of a partial analyze export
(say, if the streams tab was still loading while exporting)
to be completed at a later time if it becomes available, but
any existing analyses are unlikely to have changed, since
they depend only on the area of interest (and app config).