Centralize Data Catalog Collection init
Previously the data catalog collection would be initialized
in the data catalog controller, which was appropriate since
that was the only place it would be used.

However, since now we'll need to use it in at least one more
place, namely the upcoming Monitor tab, we move it to App,
using the same pattern as for Analyze Collection.

This is a pure refactor, and does not affect BiG-CZ behavior
at all.