Call SRAT API After Subbasin GWLF-E Completes
When we have the GWLF-E results for each HUC-12,
start a task to call the SRAT API.

Currently this will just return the results of the
API (and break the UI), which will include the
loads per HUC-12 and per each HUC-12's catchments.

Soon we will sum the HUC-12 results to get the
attenuated results for the entire shape.