Add Tabs to Analyze
Adds a tab selection header to the top of the analyze sidebar,
and a placeholder for where the monitor contents will go.

Adjusts styling to better match the mockup by making the top
tab bar grey and everything else white, and accounts for it
when calculating the tab content height so that scrolling
works correctly.