Add basic Monitor tab
This matches the mockup and demos the tab switching correctly.
It currently has no behavior.

In order to not affect BiG-CZ, it is better to create a new
set of views under /monitor rather than reuse and change the
ones under /data_catalog. We will use the same models though,
and much of the functionality will be copied over, so it is
not going to be a completely new effort.

All the styling uses flexbox.