LayerPicker: Avoid re-rendering LayoutView
When re-rendering a Marionette LayoutView, it empties
all the component regions before repopulating them. The
documentation recommends almost never re-rendering a
LayoutView, preferring to render the child views instead.

Previously, we could re-render the LayoutView and it
worked correctly enough. Unfortunately, with the jQuery
upgrade, this is no longer the case.

This commit disconnects model and collection changes
from rendering, by choosing to trigger a `onShow` method
instead which is the previous `onRender`. This ensures
that items are rendered when shown for the first time,
and subsequently updated without re-rendering the view
itself.

Other minor changes are to ensure functionality in this new
paradigm. This pattern will need to be applied throughout
the application.