$().then() -> $().done()
In jQuery 3.0, .then() was changed to always be async.
This breaks some of our tests, which assume that .then()
will be executed synchronously if added to a Deferred
which was already resolved. .done() maintains this
behavior.

This should have no material change in the front-end
behavior, and ensures that tests pass.