Adjust and clarify Geocode tests and views
Previously an unused $.Deferred() would be returned from
handleSearch() and selectFirst() methods, instead of the
correct promise of the async action. This is changed and
the main async action's promise is now returned.

Some handlers were previously added to promises in tests
as IIFEs, which caused them to be executed immediately,
rather than waiting for the promise's action. Now they
are added as normal functions, and execute when triggered
by the underlying promise.

The tests for verifying error conditions now fire on .fail()
rather than .then(), precipitated by changes to the .then()
behavior in jQuery 3.0.

See https://jquery.com/upgrade-guide/3.0/#callback-exit