Mandatorily make projects public after HydroShare
When a project is exported to HydroShare, there may be a
link from HydroShare back to the project in MMW. To allow
this, we need to make the project public in MMW so that
any incoming user can see it.

After a successful export to HydroShare, we update
the project's privacy and make it unchangeable from the
MultiShareView. We also add a message exlaining that any
project exported to HydroShare will always have link
sharing enabled.