Enable map highlight adjustments
Previously they were only enabled for BiG-CZ. Now, since the
data catalog results are also visible on MMW, we enable them
in all cases. This adds a map-highlight border around the map
when the search result is not contained within the map.

Also fixes #2336 by not removing the detail highlight when
removing the active highlight.