Add Monitor to tab switcher
Previously, since there were only two options for the sidebar
in the Modeling stage, we used a `showing_analyze` boolean on
the ProjectModel to determine what was being shown. Now that
there are three options, we use a string value defaulting to
MODEL, but also capable of being set to ANALYZE and MONITOR,
which governs what is shown on the sidebar.

The hiding and showing of various regions, as well as of the
scenario and tools toolbar, are updated to listen to this new
field instead of the old one.