Adjust bar chart ticks on tab switch
Previously, the updateChart function which adjusts ticks on
bar charts to ensure they don't overlap only fired when the
browser window was resized. Although there is a listener for
a `bar-chart:refresh` event, designed to trigger during a
tab switch, this would never fire.

By firing that event purposefully when switching tabs, it
is ensured that the ticks are always updated accordingly.