Gather HydroShare export data asynchronously
Since now to gather data for each model, this can take some
time on the front-end as we cycle through each scenario.
Previously, this was done synchronously, causing the HydroShare
Export modal to be stuck for a while until the data gathering
step was complete.

By gathering it asynchronously we ensure the modal is dismissed
and the network request fired after its completion.