Add Mean Flow values to Water Quality Summary
Mean Flow values are included in the Water Quality tab of
MapShed projects, but is not included in any table, and thus
was not included in the CSVs generated for HydroShare export.

We add two extra columns to the Water Quality Summary table,
one for per year, one for per second, and include them in the
Totals row. Other rows in that table are empty in those two
columns.

See discussion for this here:
https://github.com/WikiWatershed/model-my-watershed/issues/2666#issuecomment-369305291