Refactor dropdown toggling to use model more
Previously the model was not properly used to toggle
the dropdown. We would read from the model to get the
current state, but then directly change the UI state,
updating the model only as an afterthought after all
the changes were applied. Furthermore, since the UI
was wired to respond to model changes, this final
update would cause a redundant UI change.

Now we always only read from and change the model.
Based on the model changes, the UI is updated. This
makes for a cleaner design that is more efficient and
easier to understand.