Reuse server side validation for CUAHSI
Earlier we devised a client-side validation mechanism to not
search for CUAHSI when the area of interest is too big. While
this is faster, there is duplication of the code between client
and server. This duplication may cause mismatches down the line
if we ever change the CUAHSI_MAX_SIZE_SQKM limit.

By reusing the server side validation, we ensure that the UI
uses the same API as everyone else, and the response time change
is not very big. The presentation of the error message is worse
than purely client-side though, since there is no localization
of the numbers.