Add methods to convert result to CSV
Add five methods, each which convert a given result to a specific
kind of CSV, corresponding to the table in the UI. Which of these
should be called cannot be known by the result itself, since that
relies on information stored in the scenario (runoff and quality)
and the project (TR55 and GWLF-E). Thus, the burden of deciding
that is on the caller.

The CSV exports are made to exactly match the UI tables. When they
are changed, so will these have to be to maintain consistency.