Add utility to check for IE version
Every once in a while we need to add special behavior for
Internet Explorer. While we have so far avoided this, because
these checks tend to be brittle and often need to be updated
as newer versions are released, since Microsoft has moved on
to Edge instead of Internet Explorer, this check can safely
be used for now.

Source from: https://codepen.io/gapcode/pen/vEJNZN