Redirect to HydroShare, instead of iframe, in IE
Since IE does not allow storing third party cookies, that
makes sending requests to HydroShare from within an iframe
impossible. To alleviate this, when in IE, we redirect the
entire page instead of opening an iframe.

After the redirects complete, the user will be brought back
to their account page. This last redirect is timed so that
it happens after the iframe has been closed. Thus, the auth
landing page does not need to detect if it is IE or not. It
will only be open long enough to redirect in the IE case,
and in other cases the iframe will be closed before the
redirect happens.