Ask user to link account to HydroShare in IE
In the regular case, we can open the HydroShare auth in
an iframe and trigger the export on its success. However,
since IE does not support authorization via in iframe, we
have to go through the full page refresh cycle.

Unfortunately, the auth landing page always redirects to
the user's account page, instead of whatever project the
user was previously on. If a user were redirected from a
project page to HydroShare, they would have a reasonable
expectation to return to it after success. Since we are
not doing that here, it is better to start the redirect
process from the account page itself.

Thus we direct the user to link their account to HydroShare
via the account page, and have a link to it in the modal.
They can choose to leave the project and link their account
now, to later return to their project and export again, or
simply cancel for now and return to their immediate work.