Specify geoprocessing endpoint on call
So far we only had one endpoint in the geoprocessing service:
/run. Now we have another: /multi. This requires that we now
specify which endpoint will be targeted. The `run` task will
always go to the /run endpoint.

Furthermore, the /multi endpoint does not return its results
wrapped in a `result` key like /run does. Thus, we check if
that key is present in the response, unwrap if it is found,
else just return it as is.