Outline multi operation task, caching strategy
We follow conventions where we can to keep the API consistent.
This includes decision such as using opname (even though there
is only one multi-operation configuration, namely 'mapshed',
that will be used), caching operation results separately (even
though it is possible to cache the entire results of a multi-
operation run for a shape together), and passing exceptions
on to the next task (even though this new method removes the
use of Celery groups, which are unable to handle exceptions
in a component task, leading to this design).

While we could compose the payload outside this method, passing
in a single dict with everything it needs, doing so here encodes
some of the JSON structure into the function call (shapes,
stream_lines) which is more apparent and obvious than a single
dict argument.