Cache MultiOperation results
Before sending a multi-operation request, see if there are cached
results for each operation for a shape. If so, use them instead.
If all shapes have cached results, simply return them.

If a shape has some operation results cached but not all, the
shape will be included in the geoprocessing request and have all
operations performed on it. This is because we cannot specify
a list of operations per shape, only a list of operations and
shapes, which the geoprocessing service makes a cartesian product
of.

Returned results are saved to the cache, using the same scheme
as used by `run`. This allows for single `run` operations to reuse
cached results from a `multi` run. `multi` will reuse the cached
results if all operations for a shape are cached.

For non-well-known shapes, we pass in a special NOWKAOI shape id,
which is then used as an indicator not to cache values.