Use multi to run MapShed
A multi_mapshed task is added to gather and trigger the multi
geoprocessing operation for the given aoi or wkaoi. A convert_data
task is added to transform the results from the raw response to
the array of converted results expected by collect_data.

The view switches from using a Celery chord and a chain to a
simple chain, with the error handling applied to the entire chain
instead of individual tasks.

convert_data is given the wkaoi so it can pick which shape to
choose from the results, since it works for only a single shape.
It also logs the wkaoi in case of error, which should make traces
a little more informative.