Use multi to run Sub-basin
Add multi_subbasin task to gather streams for the parent
area of interest, and a list of child shapes, to run the
geoprocessing using the new multi endpoint.

Also add a collect_subbasins task that takes the output
of multi_subbasin and returns an array of GMS dicts by
running convert_data and collect_data for each of the
child shapes.

By removing the groups from this process, this is now a
simple Celery chain, just as done for MapShed. While
running convert_data and collect_data as a group of
chains may have been more efficient, because of a bug
in Celery we cannot have a workflow like this:

  chain(
  	task0
  	group(
  	  chain(task1, task2),
  	  chain(task3, task4))
  	task5
  	task6
  )

which gives the output of task0 to the first task
in each chain in the group, namely task1 and task3.
See https://github.com/celery/celery/issues/4244

This reuses all the code built-up for MapShed and
multi-operation, and does not change the front-end's
API at all.