Set Request Timeout On Geoprocessing Requests
If a task takes longer than the Celery HARD_TASK_TIME_LIMIT
default of 300sec, celery enters a non-responsive state. Set a timeout
on all http requests we make in tasks so we don't enter this
state while waiting for a response. This will also better alert
us when there are "silent" errors with the geoprocessing service,
ie when it is consistently timing out without other errors.