Chunk Subbasin Gwlfe Tasks
Previously we were submitting
a group where each task in the
group was `run_gwlfe` on a subbasin.
Celery waits until all these tasks
submit before allowing access to
the whole chain's job id, which is
needed for initial job submission
response. When a subbasin request
was submitted, it was taking longer
the 60s for all subbasin gwlfe tasks
to submit if there were 60+ subbasins
in a shape. To alleviate this problem,
we need to cut down the number of
tasks in the group. We do this by
chunking the subbasin into smaller
groups and then running gwlfe
synchronously inside single task
on that group.

Notes:
* Celery+redis wasn't handling
  the large GMS payloads well, and
  would hang when called with this
  chunked approach. Instead of passing
  the GMSs, read the mapshed result
  for each subbasin inside the task
  itself.

* We tried using the builtin celery
  "chunk" task, but it resulted
  in mysterious hanging not fixed
  by limiting the payload size.

* The chunk size is set to 8 based
  on local trial and error. This value
  needs to be big enough to break up
  large (60-100) sets of subbasins
  into manageble task sets, but not
  so big that the task takes longer
  than the celery hard time limit.