Add Table Tabs To Subbasin View
Subbasin view keeps its tables in tabs on the view.
Add a very simple collection to keep track of subbasin tabs,
and add collection and item views to display them.
Same as elsewhere in the app, the "Panel" views
are the tab controls and "Content" views are the
tab contents. For now the content is placeholder text.