Cache CUAHSI search results on Monitor tab select
Since CUAHSI server results don't need a text query,
and take a long time to process, we fetch them for
the entire area of interest as soon as the Monitor
tab is selected for the first time.

Afterwards, we run a dummy text search on it to ensure
its result count stays 0 until the user searching for
something. If the user switches to CUAHSI before the
initial server results are cached, we do not override
their query with the dummy text.

This way, when the user searches initially, they will
see results in HydroShare. When they switch to CUAHSI,
they will see results either immediately or as soon
as possible.