Populate a Project's Subbasin Details on Client
We will be displaying the geometries and names of a shape's
HUC-12s in subbasin mode. For larger, many-huc-ed shapes it can be
cost prohibitive to store the shapes in the actual results
where they would be stored in both the Job and Scenario tables.
Instead of having them on the 'results',
call a separate endpoint and store the details (name and shape)
of a project's subbasins in a new collection on the project: 'subbasins'.

The models in the subbasins collection, SubbasinDetailModel, have
an unused attribute 'catchments' as a placeholder for when we need
to fetch the geometries and areas of each huc12's NHD+ catchments.