Switch to more performant suds-jurko
This fork is better maintained and more performant.

suds gets installed as a transitive dependency, and once installed
it always has precedence over suds-jurko in the import statements,
so it would always be used, even when suds-jurko was available.

To ensure this doesn't happen, we remove suds after installation
and reinstall suds-jurko to ensure everything is wired up correctly.