Add HUC-12 Level Subbasin View Region
Add a new region to the model output tab for displaying
the huc12 level subbasin results (ie if you view the subbasin
results and select a single huc12 to drill down into).

Many of our existing views are almost identical to the huc12
level ones -- abstract the data-getting pieces from the views
and extend them in the HUC-12 case:
eg The subbasin SourceTableView can be reused for the huc12
level source results if the 'result' being fed in is changed to the
specific HUC-12 object. Add a 'getResult' method to the view for
Huc12SourceTableView to implement.