Improve CUAHSI search interaction
Previously, since we would trigger a CUAHSI server search
as soon as Monitor was clicked (see ca62880 and #2778),
but the text search functionality was not changed, if the
user would click CUAHSI tab before the server results were
fetched, the text search would be performed on the currently
empty set of server results and show "No Results".

Now, if a user clicks the CUAHSI tab before the server fetch
is complete, we wait for it by showing a spinner, and only
performing the text search after they are ready.

If the user does not click on CUAHSI, the behavior is same as
before.

This quirk becomes more pronounced as support for larger areas
of interest is added, causing CUAHSI to take longer to load.