Allow MapShed to run for AoIs without streams
If a MapShed run is started for an AoI that does not contain streams,
the stream related geoprocessing operation is removed from the list of
jobs. Variables and functions that use the output of that job are
modified in to account for the output not existing. This mostly involves
setting certain values to zero.

Refs #2767