Upgrade Omgeo to correctly resolve suds-jurko
Previously, python-omgeo would bring in suds as a dependency
requiring we remove it and reinstall it. This led to various
build issues.

By upgrading python-omgeo, we make it so that suds is never
brought in, and doesn't need to be uninstalled for suds-jurko
to work correctly. This fixes the build issues.