Increase front-end timeout to 160s
Previously the timeout was set to 45s. This was problematic
because the front-end timeout didn't help anyone. It doesn't
help the user, because it is giving up for them. If the user
wants to give up, they can do that. If they want to wait,
they should be able to do that too.

Furthermore, it doesn't help the server, since the timeout
doesn't actually stop the Celery job from running and using
up server resources. Even when the front end has given up,
we keep the job running.

Increasing the front end timeout from 45s to 160s for all use
cases is a simple, harmless adjustment. It will not increase
our server loads at all, and allow patient users to see results
(subbasin and otherwise) that they didn't before.

---

160s is really an unreasonable time to expect users to wait on
the internet, especially if they don't appreciate the processing
happening on the servers. We need to improve the UI for waiting
as a result of this.