Fix GMS Download Buttons
Updating the fetchGisData to use the mapshed_job_uuid instead
of the model input broke our GMS download buttons.

In addition to referencing parts of the return value of
`fetchGisData` that no longer exist, there was the added bug
that we have to apply modiciations when downloading the GMS
for a scenario, but the mod application was moved server side.

Mod application is straighforward enough to duplicate its code on
the client side. We now correctly read the gis_data from the Project
and apply the modifications from the scenario before exporting.