Keep Track Of Unapplied Gwlfe Modifications
Modifications should be applied to the GWLF-E model such
that if the user is in subbasin mode, the mods will only run
subbasin on the scenario. When/If the user exits subbasin mode,
the mods will run on the unattenuated result.

To accomplish this, add state to the scenario model to keep
track of the subbasin mode being active. When switching this
mode, re-fectch results if necessary.