Remove extra setting of HUC-12s on map
At some point, calling .showHuc12() was necessary to get the
changes / catchments on to the map. At some point, this was
switched to just listening to the 'active' state, which is
triggered by the .setActive() a line before. By doing both,
the layers in the map would be replaced by a new, identical
set, but the hover listeners would still be attached to the
old ones. This caused console errors when hovering over the
map after clicking through to the HUC-12 view by selecting
a row.

By removing the extraneous call, we ensure that there are
no console errors and hover behavior works as expected in
all cases.

Refs #2791