Add Layer Selector to Catchment Sidebar
When in the subbasin catchment view, show a layer selector that
allows the user to choose a loading rate to highlight, one of
Total Nitrogen, Total Phosphorous, and Total Sediments. The
selector looks identical to the multiple related layer selector
in Analyze.

The selected load is set on the subbasin result model, to be read
by other items in next stages.