Add Map Info region to Base Template
This will be used to hold informational popups over the map,
of which the hotspot info is the first one.

I tried to put the hotspot info within the subbasin result
template, but all its contents are restricted to within the
sidebar (overflow: hidden) so it could not be moved to the
map from there.